requirements.txt
streamlit==1.28.1
chromadb==0.4.15
langchain==0.0.340
langchain-community==0.0.1
ollama==0.1.7
pypdf2==3.0.1
pydantic==2.5.0
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
torch>=2.0.0
tiktoken==0.5.2
